chore(#48): Renamed `Entities.ts` to `entity.ts`
