+++ workspaces/backend/gql/src/enums/entity.ts
